#2674 Renamed file to reflect what it is used for
