+++ backend/census_historical_migration/management/commands/historic_workbook_generator.py
@@ -1,4 +1,4 @@
-from census_historical_migration.workbooklib.excel_creation import (
+from census_historical_migration.workbooklib.excel_creation_utils import (
     set_range,
 )
 from census_historical_migration.workbooklib.workbook_creation import (
