+++ backend/census_historical_migration/sac_general_lib/general_information.py
@@ -3,7 +3,7 @@ from datetime import date, timedelta
 from census_historical_migration.workbooklib.census_models.census import (
     CensusGen22 as Gen,
 )
-from census_historical_migration.workbooklib.excel_creation import (
+from census_historical_migration.workbooklib.excel_creation_utils import (
     _census_date_to_datetime,
 )
 from census_historical_migration.base_field_maps import FormFieldMap, FormFieldInDissem
