+++ backend/census_historical_migration/test_excel_creation.py
@@ -1,4 +1,4 @@
-from census_historical_migration.workbooklib.excel_creation import (
+from census_historical_migration.workbooklib.excel_creation_utils import (
     set_range,
 )
