+++ backend/census_historical_migration/workbooklib/additional_eins.py
@@ -1,10 +1,12 @@
-from census_historical_migration.workbooklib.excel_creation import (
-    WorkbookFieldInDissem,
+from census_historical_migration.workbooklib.excel_creation_utils import (
     set_uei,
     map_simple_columns,
     generate_dissemination_test_table,
 )
-from census_historical_migration.base_field_maps import SheetFieldMap
+from census_historical_migration.base_field_maps import (
+    SheetFieldMap,
+    WorkbookFieldInDissem,
+)
 from census_historical_migration.workbooklib.templates import sections_to_template_paths
 from census_historical_migration.workbooklib.census_models.census import dynamic_import
 from audit.fixtures.excel import FORM_SECTIONS
