+++ backend/census_historical_migration/workbooklib/secondary_auditors.py
@@ -1,9 +1,9 @@
-from census_historical_migration.workbooklib.excel_creation import (
-    SheetFieldMap,
+from census_historical_migration.workbooklib.excel_creation_utils import (
     set_uei,
     map_simple_columns,
     generate_dissemination_test_table,
 )
+from census_historical_migration.base_field_maps import SheetFieldMap
 from census_historical_migration.workbooklib.templates import sections_to_template_paths
 from census_historical_migration.workbooklib.census_models.census import dynamic_import
 from census_historical_migration.workbooklib.sac_creation import add_hyphen_to_zip
